Fix PHP 7 support for tests
