+++ .travis.yml
@@ -20,7 +20,6 @@ matrix:
         - php: 5.6
           env: DEPENDENCIES='dev'
     allow_failures:
-        - php: 7.0
         - env: DEPENDENCIES='dev'
     fast_finish: true
